+++ extensions/brew/src/search.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
-import { showFailureToast } from "./utils";
-import { Cask, Formula, InstallableResults } from "./brew";
-import { brewSearch, brewFetchInstalled } from "./brew";
+import { brewFetchInstalled, brewSearch, Cask, Formula, InstallableResults } from "./brew";
+import { InstallableFilterDropdown, InstallableFilterType } from "./components/filter";
 import { FormulaList } from "./components/list";
+import { showFailureToast } from "./utils";
 
 /// Main
 
@@ -18,10 +18,11 @@ interface State {
   results?: InstallableResults;
   installed?: Installed;
   query?: string;
+  filter: InstallableFilterType;
 }
 
 export default function Main(): JSX.Element {
-  const [state, setState] = useState<State>({ isLoading: true });
+  const [state, setState] = useState<State>({ isLoading: true, filter: InstallableFilterType.all });
 
   useEffect(() => {
     if (!state.isLoading) {
@@ -52,14 +53,23 @@ export default function Main(): JSX.Element {
       });
   }, [state]);
 
-  const formulae = state.results?.formulae ?? [];
-  const casks = state.results?.casks ?? [];
+  const formulae = state.filter != InstallableFilterType.casks ? state.results?.formulae ?? [] : [];
+  const casks = state.filter != InstallableFilterType.formulae ? state.results?.casks ?? [] : [];
 
   return (
     <FormulaList
       formulae={formulae}
       casks={casks}
       searchBarPlaceholder={"Search formulae by name" + String.ellipsis}
+      searchBarAccessory={
+        <InstallableFilterDropdown
+          onSelect={(filterType) => {
+            if (state.filter != filterType) {
+              setState((oldState) => ({ ...oldState, filter: filterType }));
+            }
+          }}
+        />
+      }
       isLoading={state.isLoading}
       onSearchTextChange={(query: string) => {
         // Perhaps query should be another useState??