+++ extensions/docsearch/src/tailwindcss3.tsx
@@ -1,0 +1,5 @@
+import { SearchDocumentation } from "./components";
+
+export default function Command() {
+  return <SearchDocumentation docsName="TailwindCSS v3" />;
+}