+++ extensions/gitlab/src/cache.ts
@@ -136,7 +136,7 @@ export function useCache<T>(
   const [isLoading, setIsLoading] = useState<boolean>(true);
   const [timestamp, setTimestamp] = useState<Date>(new Date());
   const shouldForceRefetch = useRef(false);
-  const depsAll = [timestamp];
+  const depsAll: unknown[] = [timestamp];
   if (options.deps) {
     for (const d of options.deps) {
       depsAll.push(d);