+++ extensions/linear/src/helpers/notifications.ts
@@ -3,7 +3,7 @@ import emojis from "node-emoji";
 import { NotificationResult } from "../api/getNotifications";
 import { StateType, statusIcons } from "./states";
 
-enum NotificationType {
+enum IssueNotificationType {
   issueAssignedToYou = "issueAssignedToYou",
   issueCreated = "issueCreated",
   issuePriorityUrgent = "issuePriorityUrgent",
@@ -18,60 +18,73 @@ enum NotificationType {
   issueSubscribed = "issueSubscribed",
 }
 
+enum ProjectNotificationType {
+  projectUpdatePrompt = "projectUpdatePrompt",
+  projectUpdateMentionPrompt = "projectUpdateMentionPrompt",
+}
+
 const notificationIcons: Record<string, Image.ImageLike> = {
-  [NotificationType.issueAssignedToYou]: Icon.AddPerson,
-  [NotificationType.issueCreated]: Icon.Plus,
-  [NotificationType.issuePriorityUrgent]: {
+  [IssueNotificationType.issueAssignedToYou]: Icon.AddPerson,
+  [IssueNotificationType.issueCreated]: Icon.Plus,
+  [IssueNotificationType.issuePriorityUrgent]: {
     source: { light: "light/priority-urgent.svg", dark: "dark/priority-urgent.svg" },
   },
-  [NotificationType.issueBlocking]: Icon.ExclamationMark,
-  [NotificationType.issueUnblocked]: Icon.ExclamationMark,
-  [NotificationType.issueNewComment]: Icon.Bubble,
-  [NotificationType.issueCommentMention]: Icon.Bubble,
-  [NotificationType.issueMention]: Icon.Bubble,
-  [NotificationType.issueDue]: Icon.Calendar,
-  [NotificationType.issueSubscribed]: Icon.Bell,
+  [IssueNotificationType.issueBlocking]: Icon.ExclamationMark,
+  [IssueNotificationType.issueUnblocked]: Icon.ExclamationMark,
+  [IssueNotificationType.issueNewComment]: Icon.Bubble,
+  [IssueNotificationType.issueCommentMention]: Icon.Bubble,
+  [IssueNotificationType.issueMention]: Icon.Bubble,
+  [IssueNotificationType.issueDue]: Icon.Calendar,
+  [IssueNotificationType.issueSubscribed]: Icon.Bell,
+  [ProjectNotificationType.projectUpdatePrompt]: Icon.Heartbeat,
+  [ProjectNotificationType.projectUpdateMentionPrompt]: Icon.Bubble,
 };
 
 export function getNotificationIcon(notification: NotificationResult) {
-  if (notification.type === NotificationType.issueStatusChanged) {
-    return (
-      { source: statusIcons[notification.issue.state.type], tintColor: notification.issue.state.color } || Icon.Pencil
-    );
-  }
+  if (notification.issue) {
+    if (notification.type === IssueNotificationType.issueStatusChanged) {
+      return (
+        { source: statusIcons[notification.issue.state.type], tintColor: notification.issue.state.color } || Icon.Pencil
+      );
+    }
 
-  if (notification.type === NotificationType.issueCommentReaction && notification.reactionEmoji) {
-    return emojis.get(notification.reactionEmoji) || Icon.Bubble;
+    if (notification.type === IssueNotificationType.issueCommentReaction && notification.reactionEmoji) {
+      return emojis.get(notification.reactionEmoji) || Icon.Bubble;
+    }
   }
 
   return notificationIcons[notification.type];
 }
 
 const notificationTitles: Record<string, string> = {
-  [NotificationType.issueAssignedToYou]: "Assigned to you",
-  [NotificationType.issueCreated]: "New issue created",
-  [NotificationType.issuePriorityUrgent]: "Marked as urgent",
-  [NotificationType.issueBlocking]: "Marked as blocking",
-  [NotificationType.issueUnblocked]: "Marked as unblocked",
-  [NotificationType.issueNewComment]: "New comment",
-  [NotificationType.issueCommentMention]: "Mentioned in a comment",
-  [NotificationType.issueCommentReaction]: "Reaction to a comment",
-  [NotificationType.issueMention]: "Mentioned in the issue's description",
-  [NotificationType.issueDue]: "Due soon, due, or overdue",
-  [NotificationType.issueSubscribed]: "Subscribed to the issue",
+  [IssueNotificationType.issueAssignedToYou]: "Assigned to you",
+  [IssueNotificationType.issueCreated]: "New issue created",
+  [IssueNotificationType.issuePriorityUrgent]: "Marked as urgent",
+  [IssueNotificationType.issueBlocking]: "Marked as blocking",
+  [IssueNotificationType.issueUnblocked]: "Marked as unblocked",
+  [IssueNotificationType.issueNewComment]: "New comment",
+  [IssueNotificationType.issueCommentMention]: "Mentioned in a comment",
+  [IssueNotificationType.issueCommentReaction]: "Reaction to a comment",
+  [IssueNotificationType.issueMention]: "Mentioned in the issue's description",
+  [IssueNotificationType.issueDue]: "Due soon, due, or overdue",
+  [IssueNotificationType.issueSubscribed]: "Subscribed to the issue",
+  [ProjectNotificationType.projectUpdatePrompt]: "Reminded to provide a project update",
+  [ProjectNotificationType.projectUpdateMentionPrompt]: "Mentioned in a project update",
 };
 
 export function getNotificationTitle(notification: NotificationResult) {
-  if (notification.type === NotificationType.issueStatusChanged) {
-    return (
-      {
-        [StateType.completed]: "Marked as completed",
-        [StateType.canceled]: "Marked as canceled",
-      }[notification.issue.state.type] || "Status changed"
-    );
+  if (notification.issue) {
+    if (notification.type === IssueNotificationType.issueStatusChanged) {
+      return (
+        {
+          [StateType.completed]: "Marked as completed",
+          [StateType.canceled]: "Marked as canceled",
+        }[notification.issue.state.type] || "Status changed"
+      );
+    }
   }
 
-  return notificationTitles[notification.type];
+  return notificationTitles[notification.type] || "Unknown notification";
 }
 
 export function getNotificationMenuBarIcon(unreadNotifications: NotificationResult[]) {