+++ extensions/obsidian/src/pinnedNotesCommand.tsx
@@ -2,12 +2,12 @@ import { List } from "@raycast/api";
 
 import { useObsidianVaults } from "./utils/utils";
 import { VaultSelection } from "./components/VaultSelection";
-import { Vault } from "./utils/interfaces";
+import { SearchArguments, Vault } from "./utils/interfaces";
 import { NoteListPinned } from "./components/NoteListPinned";
 import { NoVaultFoundMessage } from "./components/NoVaultFoundMessage";
 import { noVaultPathsToast } from "./components/Toasts";
 
-export default function Command() {
+export default function Command(props: { arguments: SearchArguments }) {
   const { vaults, ready } = useObsidianVaults();
 
   if (!ready) {
@@ -16,10 +16,13 @@ export default function Command() {
     return <NoVaultFoundMessage />;
   } else if (vaults.length > 1) {
     return (
-      <VaultSelection vaults={vaults} target={(vault: Vault) => <NoteListPinned vault={vault} showTitle={true} />} />
+      <VaultSelection
+        vaults={vaults}
+        target={(vault: Vault) => <NoteListPinned vault={vault} showTitle={true} searchArguments={props.arguments} />}
+      />
     );
   } else if (vaults.length == 1) {
-    return <NoteListPinned vault={vaults[0]} showTitle={false} />;
+    return <NoteListPinned vault={vaults[0]} showTitle={false} searchArguments={props.arguments} />;
   } else {
     noVaultPathsToast();
     return <List />;