+++ extensions/obsidian/src/utils/interfaces.tsx
@@ -1,3 +1,5 @@
+import { Image } from "@raycast/api";
+
 export interface FormValue {
   path: string;
   name: string;
@@ -40,6 +42,11 @@ export interface SearchNotePreferences extends GlobalPreferences, QuickLookPrefe
 
 export interface RandomNotePreferences extends GlobalPreferences, QuickLookPreferences, AppendNotePreferences {}
 
+export interface SearchMediaPreferences extends GlobalPreferences {
+  imageSize: string;
+  excludedFolders: string;
+}
+
 export interface Vault {
   name: string;
   key: string;
@@ -71,4 +78,25 @@ export interface ObsidianVaultsState {
 export interface PinnedNotesJSON {
   vaultPath: string;
   pinnedNotes: string[];
+}
+
+export interface SearchArguments {
+  searchArgument: string;
+  tagArgument: string;
+}
+
+export interface Media {
+  title: string;
+  path: string;
+  icon: Image;
+}
+
+export interface MediaState {
+  ready: boolean;
+  media: Media[];
+}
+
+export interface MediaSearchArguments {
+  searchArgument: string;
+  typeArgument: string;
 }