+++ extensions/pipe-commands/assets/commands/text/count-characters.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+# @raycast.schemaVersion 1
+# @raycast.title Count Characters
+# @raycast.mode silent
+# @raycast.icon 🔢
+# @raycast.argument1 {"type": "text", "percentEncoded": false}
+
+wc -c