+++ extensions/sourcegraph/src/sourcegraph/stream-search/stream.ts
@@ -14,6 +14,9 @@ export interface ErrorLike {
 // and will therefore default to V1.
 export const LATEST_VERSION = "V2";
 
+/** All values that are valid for the `type:` filter. `null` represents default code search. */
+export type SearchType = "file" | "repo" | "path" | "symbol" | "diff" | "commit" | null;
+
 export type SearchEvent =
   | { type: "matches"; data: SearchMatch[] }
   | { type: "progress"; data: Progress }
@@ -113,6 +116,7 @@ export interface CommitMatch {
   repoLastFetched?: string;
 
   content: MarkdownText;
+  // Array of [line, character, length] triplets
   ranges: number[][];
 }
 
@@ -208,12 +212,15 @@ export interface Filter {
   label: string;
   count: number;
   limitHit: boolean;
-  kind: string;
+  kind: "file" | "repo" | "lang" | "utility";
 }
+
+export type AlertKind = "lucky-search-queries";
 
 interface Alert {
   title: string;
   description?: string | null;
+  kind?: AlertKind | null;
   proposedQueries: ProposedQuery[] | null;
 }
 
@@ -223,6 +230,8 @@ interface ProposedQuery {
 }
 
 export type StreamingResultsState = "loading" | "error" | "complete";
+
+// Copied from the end of https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/client/shared/src/search/stream.ts?L12&subtree=true
 
 export function getRepositoryUrl(repository: string, branches?: string[]): string {
   const branch = branches?.[0];