+++ extensions/spotify-player/src/playSimilar.ts
@@ -1,18 +1,26 @@
 import { showHUD, showToast, Toast } from "@raycast/api";
-import { currentPlayingTrack, startPlaySimilar } from "./client/client";
+import { getTrack } from "./spotify/applescript";
+import { startPlaySimilar } from "./spotify/client";
+import { isAuthorized } from "./spotify/oauth";
 
 export default async function main() {
-  const response = await currentPlayingTrack();
+  const authorized = await isAuthorized();
+  if (!authorized) {
+    showHUD("⚠️ Please open any view-based command and authorize to perform the command.");
+    return;
+  }
+  try {
+    const track = await getTrack();
 
-  if (response?.result) {
-    const currentlyPlayingTrack = response.result;
-    const trackTitle = `${currentlyPlayingTrack.artist} – ${currentlyPlayingTrack.name}`;
-    if (currentlyPlayingTrack.id) {
-      const trackId = currentlyPlayingTrack.id.replace("spotify:track:", "");
-      await startPlaySimilar(trackId);
-      showHUD(`♫ Playing Similar – ♫ ${trackTitle}`);
+    if (track) {
+      const trackTitle = `${track.artist} – ${track.name}`;
+      if (track.id) {
+        const trackId = track.id.replace("spotify:track:", "");
+        await startPlaySimilar({ seed_tracks: trackId });
+        showHUD(`♫ Playing Similar – ♫ ${trackTitle}`);
+      }
     }
-  } else if (response?.error) {
-    showToast(Toast.Style.Failure, response.error);
+  } catch (error) {
+    showToast(Toast.Style.Failure, String(error));
   }
 }