+++ extensions/xcode/src/services/xcode.service.ts
@@ -5,9 +5,16 @@ import { execAsync } from "../shared/exec-async";
  */
 export class XcodeService {
   /**
+   * Xcode application bundle identifier
+   */
+  static get bundleIdentifier(): string {
+    return "com.apple.dt.Xcode";
+  }
+
+  /**
    * Retrieve boolean if Xcode is currently running
    */
-  async isXcodeRunning(): Promise<boolean> {
+  static async isXcodeRunning(): Promise<boolean> {
     try {
       // prep Xcode process status
       const output = await execAsync("pgrep Xcode");
@@ -22,12 +29,12 @@ export class XcodeService {
   /**
    * Launch Xcode
    */
-  launchXcode(): Promise<void> {
+  static launchXcode(): Promise<void> {
     return execAsync(
       [
         // Open Xcode in background (-j)
         // via bundle identifier (-b)
-        "open -j -b com.apple.dt.xcode",
+        `open -j -b ${XcodeService.bundleIdentifier}`,
         // Sleep for two seconds to ensure
         // the process is truly running
         "sleep 2",