+++ extensions/xcode/src/show-recent-projects-in-menu-bar.command.tsx
@@ -1,0 +1,5 @@
+import { XcodeProjectsMenuBar } from "./components/xcode-projects-menu-bar/xcode-projects-menu-bar.component";
+
+export default () => {
+  return <XcodeProjectsMenuBar />;
+};